Move Statistics class to separate file
