+++ pynicotine/transfers.py
@@ -2532,43 +2532,3 @@ class Transfers:
 
     def quit(self):
         self.save_transfers()
-
-
-class Statistics:
-
-    def __init__(self, ui_callback=None):
-
-        self.ui_callback = getattr(ui_callback, "statistics", None)
-        self.session_stats = {}
-
-        # Only populate total since date on first run
-        if (not config.sections["statistics"]["since_timestamp"]
-                and config.sections["statistics"] == config.defaults["statistics"]):
-            config.sections["statistics"]["since_timestamp"] = int(time.time())
-
-        for stat_id in config.defaults["statistics"]:
-            self.session_stats[stat_id] = 0 if stat_id != "since_timestamp" else int(time.time())
-
-        for stat_id in config.defaults["statistics"]:
-            self.update_ui(stat_id)
-
-    def append_stat_value(self, stat_id, stat_value):
-
-        self.session_stats[stat_id] += stat_value
-        config.sections["statistics"][stat_id] += stat_value
-        self.update_ui(stat_id)
-
-    def update_ui(self, stat_id):
-
-        if self.ui_callback:
-            session_stat_value = self.session_stats[stat_id]
-            total_stat_value = config.sections["statistics"][stat_id]
-            self.ui_callback.update_stat_value(stat_id, session_stat_value, total_stat_value)
-
-    def reset_stats(self):
-
-        for stat_id in config.defaults["statistics"]:
-            stat_value = 0 if stat_id != "since_timestamp" else int(time.time())
-            self.session_stats[stat_id] = config.sections["statistics"][stat_id] = stat_value
-
-            self.update_ui(stat_id)
